+++ platform/darwin/src/native_apple_interface.m
@@ -53,6 +53,7 @@ static MGLNativeNetworkManager *instance = nil;
 }
 
 - (void)stopDownloadEventForResponse:(NSURLResponse *)response {
+
     if (_delegate && [_delegate respondsToSelector:@selector(stopDownloadEventForResponse:)]) {
         return [_delegate stopDownloadEventForResponse:response];
     }