+++ render-test/ios/ViewController.h
@@ -1,0 +1,9 @@
+// ViewController.h
+// Your main view controller
+
+#import <UIKit/UIKit.h>
+
+@interface ViewController : UIViewController
+
+@end
+